Completed 4.21-2.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog') 
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const { userExtractor } = require("../utils/middleware");
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -9,15 +9,10 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const { title, author, ...body } = request.body
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
+  const user = request.user
 
   if (!title || !author) {
     return response.status(400).json({ error: 'title or author missing' })
@@ -37,11 +32,17 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
     const blog = await Blog.findById(request.params.id)
 
     if(!blog) {
         return response.status(400).json({ error: 'invalid id' })
+    }
+
+    const user = request.user
+
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({ error: 'unauthorized to delete' })
     }
 
     await Blog.findByIdAndDelete(request.params.id)
